+++ demo/nextjs/pages/core-initial-page/index.tsx
@@ -1,0 +1,20 @@
+import * as React from 'react';
+import { Viewer } from '@react-pdf-viewer/core';
+
+const IndexPage = () => {
+    return (
+        <div
+            style={{
+                border: '1px solid rgba(0, 0, 0, .3)',
+                display: 'flex',
+                height: '50rem',
+                margin: '5rem auto',
+                width: '64rem',
+            }}
+        >
+            <Viewer fileUrl="/pdf-open-parameters.pdf" initialPage={6} />
+        </div>
+    );
+};
+
+export default IndexPage;